NOTE: THE MODEL STILL DOES NOT LEARN CORRECTLY.
Modified the DeepLabv3 classifier in order to allow us to use pre-trained weights with our input/output.

Setup the preprocessing of the images to be the same as those used in the pre-training.

Reworked the visualize_results function to be able to optionally visualize the input and masks as they were used in the training/evaluation instead of always re-loading the images fresh.

Commented out code that adapts the model to accept 6-channel input. This was done simply to try and get the model to work with a standard 3-channel RGB image and a corresponding mask.