Created a script that uses rasterio reprojection to align one image geographically against another and then cuts the images into equal slices.
Adding the new dataset with unusable images removed for pre and post images, masks, elevation, hand, and imperviousness. Removed old dataset files.

Added working data augmentation to the dataloader.

Reworked the evaluation metrics to report scores per class instead of the overall average.

Updated the Notebook version of the code to the latest.